+++ project/3rdparty/include/xlib/xlib/xlibconfig.h
@@ -138,7 +138,7 @@ typedef unsigned int xuintptr;
 
 #define XLIB_MAJOR_VERSION                      2
 #define XLIB_MINOR_VERSION                      81
-#define XLIB_MICRO_VERSION                      0
+#define XLIB_MICRO_VERSION                      2
 
 #define X_OS_UNIX