+++ project/3rdparty/source/xlib/xlib/xthread-posix.c
@@ -31,12 +31,110 @@ static void x_thread_abort(xint status, const xchar *function)
     x_abort();
 }
 
+#if !defined(USE_NATIVE_MUTEX)
+
+static pthread_mutex_t *x_mutex_impl_new(void)
+{
+    xint status;
+    pthread_mutex_t *mutex;
+    pthread_mutexattr_t *pattr = NULL;
+#ifdef PTHREAD_ADAPTIVE_MUTEX_INITIALIZER_NP
+    pthread_mutexattr_t attr;
+#endif
+
+    mutex = malloc(sizeof(pthread_mutex_t));
+    if X_UNLIKELY(mutex == NULL) {
+        x_thread_abort(errno, "malloc");
+    }
+
+#ifdef PTHREAD_ADAPTIVE_MUTEX_INITIALIZER_NP
+    pthread_mutexattr_init (&attr);
+    pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_ADAPTIVE_NP);
+    pattr = &attr;
+#endif
+
+    if X_UNLIKELY((status = pthread_mutex_init(mutex, pattr)) != 0) {
+        x_thread_abort(status, "pthread_mutex_init");
+    }
+
+#ifdef PTHREAD_ADAPTIVE_MUTEX_INITIALIZER_NP
+    pthread_mutexattr_destroy(&attr);
+#endif
+
+    return mutex;
+}
+
+static void x_mutex_impl_free(pthread_mutex_t *mutex)
+{
+    pthread_mutex_destroy(mutex);
+    free(mutex);
+}
+
+static inline pthread_mutex_t *x_mutex_get_impl(XMutex *mutex)
+{
+    pthread_mutex_t *impl = x_atomic_pointer_get(&mutex->p);
+
+    if X_UNLIKELY(impl == NULL) {
+        impl = x_mutex_impl_new();
+        if (!x_atomic_pointer_compare_and_exchange(&mutex->p, NULL, impl)) {
+            x_mutex_impl_free(impl);
+        }
+        impl = mutex->p;
+    }
+
+    return impl;
+}
+
+void x_mutex_init_impl(XMutex *mutex)
+{
+    mutex->p = x_mutex_impl_new();
+}
+
+void x_mutex_clear_impl(XMutex *mutex)
+{
+    x_mutex_impl_free(mutex->p);
+}
+
+void x_mutex_lock_impl(XMutex *mutex)
+{
+    xint status;
+
+    if X_UNLIKELY((status = pthread_mutex_lock(x_mutex_get_impl(mutex))) != 0) {
+        x_thread_abort(status, "pthread_mutex_lock");
+    }
+}
+
+void x_mutex_unlock_impl(XMutex *mutex)
+{
+    xint status;
+
+    if X_UNLIKELY((status = pthread_mutex_unlock(x_mutex_get_impl(mutex))) != 0) {
+        x_thread_abort(status, "pthread_mutex_unlock");
+    }
+}
+
+xboolean x_mutex_trylock_impl(XMutex *mutex)
+{
+    xint status;
+
+    if X_LIKELY((status = pthread_mutex_trylock(x_mutex_get_impl(mutex))) == 0) {
+        return TRUE;
+    }
+
+    if X_UNLIKELY(status != EBUSY) {
+        x_thread_abort(status, "pthread_mutex_trylock");
+    }
+
+    return FALSE;
+}
+#endif
+
 static pthread_mutex_t *x_rec_mutex_impl_new(void)
 {
     pthread_mutex_t *mutex;
     pthread_mutexattr_t attr;
 
-    mutex = (pthread_mutex_t *)malloc(sizeof(pthread_mutex_t));
+    mutex = malloc(sizeof(pthread_mutex_t));
     if X_UNLIKELY(mutex == NULL) {
         x_thread_abort(errno, "malloc");
     }
@@ -57,7 +155,7 @@ static void x_rec_mutex_impl_free(pthread_mutex_t *mutex)
 
 static inline pthread_mutex_t *x_rec_mutex_get_impl(XRecMutex *rec_mutex)
 {
-    pthread_mutex_t *impl = (pthread_mutex_t *)x_atomic_pointer_get(&rec_mutex->p);
+    pthread_mutex_t *impl = x_atomic_pointer_get(&rec_mutex->p);
 
     if X_UNLIKELY(impl == NULL) {
         impl = x_rec_mutex_impl_new();
@@ -65,33 +163,33 @@ static inline pthread_mutex_t *x_rec_mutex_get_impl(XRecMutex *rec_mutex)
             x_rec_mutex_impl_free(impl);
         }
 
-        impl = (pthread_mutex_t *)rec_mutex->p;
+        impl = rec_mutex->p;
     }
 
     return impl;
 }
 
-void x_rec_mutex_init(XRecMutex *rec_mutex)
+void x_rec_mutex_init_impl(XRecMutex *rec_mutex)
 {
     rec_mutex->p = x_rec_mutex_impl_new();
 }
 
-void x_rec_mutex_clear(XRecMutex *rec_mutex)
+void x_rec_mutex_clear_impl(XRecMutex *rec_mutex)
 {
-    x_rec_mutex_impl_free((pthread_mutex_t *)rec_mutex->p);
+    x_rec_mutex_impl_free(rec_mutex->p);
 }
 
-void x_rec_mutex_lock(XRecMutex *mutex)
+void x_rec_mutex_lock_impl(XRecMutex *mutex)
 {
     pthread_mutex_lock(x_rec_mutex_get_impl(mutex));
 }
 
-void x_rec_mutex_unlock(XRecMutex *rec_mutex)
+void x_rec_mutex_unlock_impl(XRecMutex *rec_mutex)
 {
-    pthread_mutex_unlock((pthread_mutex_t *)rec_mutex->p);
+    pthread_mutex_unlock(rec_mutex->p);
 }
 
-xboolean x_rec_mutex_trylock(XRecMutex *rec_mutex)
+xboolean x_rec_mutex_trylock_impl(XRecMutex *rec_mutex)
 {
     if (pthread_mutex_trylock(x_rec_mutex_get_impl(rec_mutex)) != 0) {
         return FALSE;
@@ -105,7 +203,7 @@ static pthread_rwlock_t *x_rw_lock_impl_new(void)
     xint status;
     pthread_rwlock_t *rwlock;
 
-    rwlock = (pthread_rwlock_t *)malloc(sizeof(pthread_rwlock_t));
+    rwlock = malloc(sizeof(pthread_rwlock_t));
     if X_UNLIKELY(rwlock == NULL) {
         x_thread_abort(errno, "malloc");
     }
@@ -125,40 +223,38 @@ static void x_rw_lock_impl_free(pthread_rwlock_t *rwlock)
 
 static inline pthread_rwlock_t *x_rw_lock_get_impl(XRWLock *lock)
 {
-    pthread_rwlock_t *impl = (pthread_rwlock_t *)x_atomic_pointer_get(&lock->p);
+    pthread_rwlock_t *impl = x_atomic_pointer_get(&lock->p);
 
     if X_UNLIKELY(impl == NULL) {
         impl = x_rw_lock_impl_new();
         if (!x_atomic_pointer_compare_and_exchange(&lock->p, NULL, impl)) {
             x_rw_lock_impl_free(impl);
         }
-
-        impl = (pthread_rwlock_t *)lock->p;
+        impl = lock->p;
     }
 
     return impl;
 }
 
-void x_rw_lock_init(XRWLock *rw_lock)
+void x_rw_lock_init_impl(XRWLock *rw_lock)
 {
     rw_lock->p = x_rw_lock_impl_new();
 }
 
-void x_rw_lock_clear(XRWLock *rw_lock)
+void x_rw_lock_clear_impl(XRWLock *rw_lock)
 {
-    x_rw_lock_impl_free((pthread_rwlock_t *)rw_lock->p);
+    x_rw_lock_impl_free(rw_lock->p);
 }
 
-void x_rw_lock_writer_lock(XRWLock *rw_lock)
+void x_rw_lock_writer_lock_impl(XRWLock *rw_lock)
 {
     int retval = pthread_rwlock_wrlock(x_rw_lock_get_impl(rw_lock));
-
     if (retval != 0) {
         x_critical("Failed to get RW lock %p: %s", rw_lock, x_strerror(retval));
     }
 }
 
-xboolean x_rw_lock_writer_trylock(XRWLock *rw_lock)
+xboolean x_rw_lock_writer_trylock_impl(XRWLock *rw_lock)
 {
     if (pthread_rwlock_trywrlock(x_rw_lock_get_impl(rw_lock)) != 0) {
         return FALSE;
@@ -167,20 +263,20 @@ xboolean x_rw_lock_writer_trylock(XRWLock *rw_lock)
     return TRUE;
 }
 
-void x_rw_lock_writer_unlock(XRWLock *rw_lock)
+void x_rw_lock_writer_unlock_impl(XRWLock *rw_lock)
 {
     pthread_rwlock_unlock(x_rw_lock_get_impl(rw_lock));
 }
 
-void x_rw_lock_reader_lock(XRWLock *rw_lock)
+void x_rw_lock_reader_lock_impl(XRWLock *rw_lock)
 {
-    int retval = pthread_rwlock_rdlock(x_rw_lock_get_impl(rw_lock));
+    int retval = pthread_rwlock_rdlock (x_rw_lock_get_impl(rw_lock));
     if (retval != 0) {
         x_critical("Failed to get RW lock %p: %s", rw_lock, x_strerror(retval));
     }
 }
 
-xboolean x_rw_lock_reader_trylock(XRWLock *rw_lock)
+xboolean x_rw_lock_reader_trylock_impl(XRWLock *rw_lock)
 {
     if (pthread_rwlock_tryrdlock(x_rw_lock_get_impl(rw_lock)) != 0) {
         return FALSE;
@@ -189,17 +285,150 @@ xboolean x_rw_lock_reader_trylock(XRWLock *rw_lock)
     return TRUE;
 }
 
-void x_rw_lock_reader_unlock(XRWLock *rw_lock)
+void x_rw_lock_reader_unlock_impl(XRWLock *rw_lock)
 {
     pthread_rwlock_unlock(x_rw_lock_get_impl(rw_lock));
 }
+
+#if !defined(USE_NATIVE_MUTEX)
+static pthread_cond_t *x_cond_impl_new(void)
+{
+    xint status;
+    pthread_cond_t *cond;
+    pthread_condattr_t attr;
+
+    pthread_condattr_init (&attr);
+
+#ifdef HAVE_PTHREAD_COND_TIMEDWAIT_RELATIVE_NP
+#elif defined (HAVE_PTHREAD_CONDATTR_SETCLOCK) && defined (CLOCK_MONOTONIC)
+    if X_UNLIKELY((status = pthread_condattr_setclock(&attr, CLOCK_MONOTONIC)) != 0) {
+        x_thread_abort(status, "pthread_condattr_setclock");
+    }
+#else
+    #error Cannot support XCond on your platform.
+#endif
+
+    cond = malloc(sizeof(pthread_cond_t));
+    if X_UNLIKELY(cond == NULL) {
+        x_thread_abort(errno, "malloc");
+    }
+
+    if X_UNLIKELY((status = pthread_cond_init (cond, &attr)) != 0) {
+        x_thread_abort(status, "pthread_cond_init");
+    }
+
+    pthread_condattr_destroy(&attr);
+    return cond;
+}
+
+static void x_cond_impl_free(pthread_cond_t *cond)
+{
+    pthread_cond_destroy(cond);
+    free(cond);
+}
+
+static inline pthread_cond_t *x_cond_get_impl(XCond *cond)
+{
+    pthread_cond_t *impl = x_atomic_pointer_get(&cond->p);
+
+    if X_UNLIKELY(impl == NULL) {
+        impl = x_cond_impl_new();
+        if (!x_atomic_pointer_compare_and_exchange(&cond->p, NULL, impl)) {
+            x_cond_impl_free(impl);
+        }
+        impl = cond->p;
+    }
+
+    return impl;
+}
+
+void x_cond_init_impl(XCond *cond)
+{
+    cond->p = x_cond_impl_new();
+}
+
+void x_cond_clear_impl(XCond *cond)
+{
+    x_cond_impl_free(cond->p);
+}
+
+void x_cond_wait_impl(XCond *cond, XMutex *mutex)
+{
+    xint status;
+
+    if X_UNLIKELY((status = pthread_cond_wait(x_cond_get_impl(cond), x_mutex_get_impl(mutex))) != 0) {
+        x_thread_abort(status, "pthread_cond_wait");
+    }
+}
+
+void x_cond_signal_impl(XCond *cond)
+{
+    xint status;
+
+    if X_UNLIKELY((status = pthread_cond_signal(x_cond_get_impl(cond))) != 0) {
+        x_thread_abort(status, "pthread_cond_signal");
+    }
+}
+
+void x_cond_broadcast_impl(XCond *cond)
+{
+    xint status;
+
+    if X_UNLIKELY((status = pthread_cond_broadcast(g_cond_get_impl(cond))) != 0) {
+        x_thread_abort(status, "pthread_cond_broadcast");
+    }
+}
+
+xboolean x_cond_wait_until_impl(XCond *cond, XMutex *mutex, xint64 end_time)
+{
+    xint status;
+    struct timespec ts;
+
+#ifdef HAVE_PTHREAD_COND_TIMEDWAIT_RELATIVE_NP
+    {
+        xint64 relative;
+        xint64 now = x_get_monotonic_time();
+
+        if (end_time <= now) {
+            return FALSE;
+        }
+
+        relative = end_time - now;
+
+        ts.tv_sec = relative / 1000000;
+        ts.tv_nsec = (relative % 1000000) * 1000;
+
+        if ((status = pthread_cond_timedwait_relative_np(x_cond_get_impl(cond), x_mutex_get_impl(mutex), &ts)) == 0) {
+            return TRUE;
+        }
+    }
+#elif defined (HAVE_PTHREAD_CONDATTR_SETCLOCK) && defined (CLOCK_MONOTONIC)
+    {
+        ts.tv_sec = end_time / 1000000;
+        ts.tv_nsec = (end_time % 1000000) * 1000;
+
+        if ((status = pthread_cond_timedwait(x_cond_get_impl(cond), x_mutex_get_impl(mutex), &ts)) == 0) {
+            return TRUE;
+        }
+    }
+#else
+    #error Cannot support XCond on your platform.
+#endif
+
+    if X_UNLIKELY(status != ETIMEDOUT) {
+        x_thread_abort(status, "pthread_cond_timedwait");
+    }
+
+    return FALSE;
+}
+#endif
 
 static pthread_key_t *x_private_impl_new(XDestroyNotify notify)
 {
     xint status;
     pthread_key_t *key;
 
-    key = (pthread_key_t *)malloc(sizeof(pthread_key_t));
+    key = malloc(sizeof(pthread_key_t));
     if X_UNLIKELY(key == NULL) {
         x_thread_abort(errno, "malloc");
     }
@@ -228,7 +457,7 @@ static xpointer x_private_impl_new_direct(XDestroyNotify notify)
 {
     xint status;
     pthread_key_t key;
-    xpointer impl = (void *)(xsize)-1;
+    xpointer impl = (void *)(xssize)-1;
 
     status = pthread_key_create(&key, notify);
     if X_UNLIKELY(status != 0) {
@@ -268,11 +497,10 @@ static void x_private_impl_free_direct(xpointer impl)
     }
 }
 
-static inline pthread_key_t x_private_get_impl(XPrivate *key)
+static inline pthread_key_t _x_private_get_impl(XPrivate *key)
 {
     if (sizeof(pthread_key_t) > sizeof(xpointer)) {
         pthread_key_t *impl = x_atomic_pointer_get(&key->p);
-
         if X_UNLIKELY(impl == NULL) {
             impl = x_private_impl_new(key->notify);
             if (!x_atomic_pointer_compare_and_exchange(&key->p, NULL, impl)) {
@@ -299,25 +527,25 @@ static inline pthread_key_t x_private_get_impl(XPrivate *key)
     }
 }
 
-xpointer x_private_get(XPrivate *key)
+xpointer x_private_get_impl(XPrivate *key)
 {
-    return pthread_getspecific(x_private_get_impl(key));
+    return pthread_getspecific(_x_private_get_impl(key));
 }
 
-void x_private_set(XPrivate *key, xpointer value)
+void x_private_set_impl(XPrivate *key, xpointer value)
 {
     xint status;
 
-    if X_UNLIKELY((status = pthread_setspecific(x_private_get_impl(key), value)) != 0) {
+    if X_UNLIKELY((status = pthread_setspecific (_x_private_get_impl(key), value)) != 0) {
         x_thread_abort(status, "pthread_setspecific");
     }
 }
 
-void x_private_replace(XPrivate *key, xpointer value)
+void x_private_replace_impl(XPrivate *key, xpointer value)
 {
     xint status;
     xpointer old;
-    pthread_key_t impl = x_private_get_impl(key);
+    pthread_key_t impl = _x_private_get_impl(key);
 
     old = pthread_getspecific(impl);
 
@@ -326,7 +554,7 @@ void x_private_replace(XPrivate *key, xpointer value)
     }
 
     if (old && key->notify) {
-        key->notify(old);
+        key->notify (old);
     }
 }
 
@@ -338,21 +566,19 @@ void x_private_replace(XPrivate *key, xpointer value)
         }                                                                                                                       \
     } X_STMT_END
 
-#define posix_check_cmd(cmd)        posix_check_err(cmd, #cmd)
+#define posix_check_cmd(cmd)    posix_check_err(cmd, #cmd)
 
 typedef struct {
     XRealThread thread;
     pthread_t   system_thread;
     xboolean    joined;
     XMutex      lock;
-
-    void *(*proxy) (void *);
+    void *(*proxy)(void *);
 } XThreadPosix;
 
 void x_system_thread_free(XRealThread *thread)
 {
     XThreadPosix *pt = (XThreadPosix *)thread;
-
     if (!pt->joined) {
         pthread_detach(pt->system_thread);
     }
@@ -388,33 +614,33 @@ XRealThread *x_system_thread_new(XThreadFunc proxy, xulong stack_size, const cha
             stack_size = MAX((xulong)min_stack_size, stack_size);
         }
 #endif
-        pthread_attr_setstacksize(&attr, stack_size);
+        pthread_attr_setstacksize (&attr, stack_size);
     }
 #endif
 
 #ifdef HAVE_PTHREAD_ATTR_SETINHERITSCHED
     {
-        pthread_attr_setinheritsched (&attr, PTHREAD_INHERIT_SCHED);
+        pthread_attr_setinheritsched(&attr, PTHREAD_INHERIT_SCHED);
     }
 #endif
 
     ret = pthread_create(&thread->system_thread, &attr, (void *(*)(void *))proxy, thread);
-    posix_check_cmd(pthread_attr_destroy(&attr));
+    posix_check_cmd(pthread_attr_destroy (&attr));
 
     if (ret == EAGAIN) {
-        x_set_error(error, X_THREAD_ERROR, X_THREAD_ERROR_AGAIN,  "Error creating thread: %s", x_strerror(ret));
+        x_set_error(error, X_THREAD_ERROR, X_THREAD_ERROR_AGAIN, "Error creating thread: %s", x_strerror(ret));
         x_free(thread->thread.name);
         x_slice_free(XThreadPosix, thread);
         return NULL;
     }
 
     posix_check_err(ret, "pthread_create");
-
     x_mutex_init(&thread->lock);
+
     return (XRealThread *)thread;
 }
 
-void x_thread_yield(void)
+void x_thread_yield_impl(void)
 {
     sched_yield();
 }
@@ -449,11 +675,26 @@ void x_system_thread_set_name(const xchar *name)
 #endif
 }
 
-#define exchange_acquire(ptr, newt)              __atomic_exchange_4((ptr), (newt), __ATOMIC_ACQUIRE)
-#define compare_exchange_acquire(ptr, old, newt) __atomic_compare_exchange_4((ptr), (old), (newt), 0, __ATOMIC_ACQUIRE, __ATOMIC_RELAXED)
+#if defined(USE_NATIVE_MUTEX)
+#ifdef HAVE_STDATOMIC_H
 
-#define exchange_release(ptr, newt)              __atomic_exchange_4((ptr), (newt), __ATOMIC_RELEASE)
-#define store_release(ptr, newt)                 __atomic_store_4((ptr), (newt), __ATOMIC_RELEASE)
+#include <stdatomic.h>
+
+#define exchange_acquire(ptr, new)                  atomic_exchange_explicit((atomic_uint *)(ptr), (new), __ATOMIC_ACQUIRE)
+#define compare_exchange_acquire(ptr, old, new)     atomic_compare_exchange_strong_explicit((atomic_uint *)(ptr), (old), (new), __ATOMIC_ACQUIRE, __ATOMIC_RELAXED)
+
+#define exchange_release(ptr, new)                  atomic_exchange_explicit((atomic_uint *)(ptr), (new), __ATOMIC_RELEASE)
+#define store_release(ptr, new)                     atomic_store_explicit((atomic_uint *)(ptr), (new), __ATOMIC_RELEASE)
+
+#else
+
+#define exchange_acquire(ptr, new)                  __atomic_exchange_4((ptr), (new), __ATOMIC_ACQUIRE)
+#define compare_exchange_acquire(ptr, old, new)     __atomic_compare_exchange_4((ptr), (old), (new), 0, __ATOMIC_ACQUIRE, __ATOMIC_RELAXED)
+
+#define exchange_release(ptr, new)                  __atomic_exchange_4((ptr), (new), __ATOMIC_RELEASE)
+#define store_release(ptr, new)                     __atomic_store_4((ptr), (new), __ATOMIC_RELEASE)
+
+#endif
 
 typedef enum {
     X_MUTEX_STATE_EMPTY = 0,
@@ -461,12 +702,12 @@ typedef enum {
     X_MUTEX_STATE_CONTENDED,
 } XMutexState;
 
-void x_mutex_init(XMutex *mutex)
+void x_mutex_init_impl(XMutex *mutex)
 {
     mutex->i[0] = X_MUTEX_STATE_EMPTY;
 }
 
-void x_mutex_clear(XMutex *mutex)
+void x_mutex_clear_impl(XMutex *mutex)
 {
     if X_UNLIKELY(mutex->i[0] != X_MUTEX_STATE_EMPTY) {
         fprintf(stderr, "x_mutex_clear() called on uninitialised or locked mutex\n");
@@ -474,58 +715,59 @@ void x_mutex_clear(XMutex *mutex)
     }
 }
 
-X_GNUC_NO_INLINE static void x_mutex_lock_slowpath(XMutex *mutex)
+X_GNUC_NO_INLINE
+static void x_mutex_lock_slowpath(XMutex *mutex)
 {
     while (exchange_acquire(&mutex->i[0], X_MUTEX_STATE_CONTENDED) != X_MUTEX_STATE_EMPTY) {
         x_futex_simple(&mutex->i[0], (xsize)FUTEX_WAIT_PRIVATE, X_MUTEX_STATE_CONTENDED, NULL);
     }
 }
 
-X_GNUC_NO_INLINE static void x_mutex_unlock_slowpath(XMutex *mutex, xuint   prev)
+X_GNUC_NO_INLINE
+static void x_mutex_unlock_slowpath(XMutex *mutex, xuint prev)
 {
     if X_UNLIKELY(prev == X_MUTEX_STATE_EMPTY) {
         fprintf(stderr, "Attempt to unlock mutex that was not locked\n");
         x_abort();
     }
 
-    x_futex_simple(&mutex->i[0], (xsize)FUTEX_WAKE_PRIVATE, (xsize)1, NULL);
+    x_futex_simple(&mutex->i[0], (xsize)FUTEX_WAKE_PRIVATE, (xsize) 1, NULL);
 }
 
-void x_mutex_lock(XMutex *mutex)
+void x_mutex_lock_impl(XMutex *mutex)
 {
     if X_UNLIKELY(!x_atomic_int_compare_and_exchange(&mutex->i[0], X_MUTEX_STATE_EMPTY, X_MUTEX_STATE_OWNED)) {
         x_mutex_lock_slowpath(mutex);
     }
 }
 
-void x_mutex_unlock(XMutex *mutex)
+void x_mutex_unlock_impl(XMutex *mutex)
 {
     xuint prev;
 
     prev = exchange_release(&mutex->i[0], X_MUTEX_STATE_EMPTY);
-
     if X_UNLIKELY(prev != X_MUTEX_STATE_OWNED) {
         x_mutex_unlock_slowpath(mutex, prev);
     }
 }
 
-xboolean x_mutex_trylock(XMutex *mutex)
+xboolean x_mutex_trylock_impl(XMutex *mutex)
 {
     XMutexState empty = X_MUTEX_STATE_EMPTY;
     return compare_exchange_acquire(&mutex->i[0], &empty, X_MUTEX_STATE_OWNED);
 }
 
-void x_cond_init(XCond *cond)
+void x_cond_init_impl(XCond *cond)
 {
     cond->i[0] = 0;
 }
 
-void x_cond_clear(XCond *cond)
+void x_cond_clear_impl(XCond *cond)
 {
 
 }
 
-void x_cond_wait(XCond *cond, XMutex *mutex)
+void x_cond_wait_impl(XCond *cond, XMutex *mutex)
 {
     xuint sampled = (xuint)x_atomic_int_get(&cond->i[0]);
 
@@ -534,19 +776,19 @@ void x_cond_wait(XCond *cond, XMutex *mutex)
     x_mutex_lock(mutex);
 }
 
-void x_cond_signal(XCond *cond)
+void x_cond_signal_impl(XCond *cond)
 {
     x_atomic_int_inc(&cond->i[0]);
     x_futex_simple(&cond->i[0], (xsize)FUTEX_WAKE_PRIVATE, (xsize)1, NULL);
 }
 
-void x_cond_broadcast(XCond *cond)
+void x_cond_broadcast_impl(XCond *cond)
 {
     x_atomic_int_inc(&cond->i[0]);
     x_futex_simple(&cond->i[0], (xsize)FUTEX_WAKE_PRIVATE, (xsize)INT_MAX, NULL);
 }
 
-xboolean x_cond_wait_until(XCond *cond, XMutex *mutex, xint64 end_time)
+xboolean x_cond_wait_until_impl(XCond *cond, XMutex *mutex, xint64 end_time)
 {
     int res;
     xuint sampled;
@@ -574,11 +816,7 @@ xboolean x_cond_wait_until(XCond *cond, XMutex *mutex, xint64 end_time)
     x_mutex_unlock(mutex);
 
 #if defined(HAVE_FUTEX_TIME64)
-#if defined(__ANDROID__)
-    if (__builtin_available (android 30, *)) {
-#else
     {
-#endif
         struct {
             xint64 tv_sec;
             xint64 tv_nsec;
@@ -595,7 +833,6 @@ xboolean x_cond_wait_until(XCond *cond, XMutex *mutex, xint64 end_time)
         {
             success = (res < 0 && errno == ETIMEDOUT) ? FALSE : TRUE;
             x_mutex_lock(mutex);
-
             return success;
         }
     }
@@ -604,19 +841,19 @@ xboolean x_cond_wait_until(XCond *cond, XMutex *mutex, xint64 end_time)
 #if defined(HAVE_FUTEX)
     {
 #ifdef __kernel_long_t
-#define KERNEL_SPAN_SEC_TYPE __kernel_long_t
+#define KERNEL_SPAN_SEC_TYPE    __kernel_long_t
         struct {
-            __kernel_long_t tv_sec;
-            __kernel_long_t tv_nsec;
+        __kernel_long_t tv_sec;
+        __kernel_long_t tv_nsec;
         } span_arg;
 #else
-#define KERNEL_SPAN_SEC_TYPE __kernel_time_t
-    struct {
-      __kernel_time_t tv_sec;
-      long            tv_nsec;
-    } span_arg;
+#define KERNEL_SPAN_SEC_TYPE    __kernel_time_t
+        struct {
+            __kernel_time_t tv_sec;
+            long            tv_nsec;
+        } span_arg;
 #endif
-        if (X_UNLIKELY(sizeof(KERNEL_SPAN_SEC_TYPE) < 8 && span.tv_sec > X_MAXINT32)) {
+        if (X_UNLIKELY((sizeof(KERNEL_SPAN_SEC_TYPE) < 8) && (span.tv_sec > X_MAXINT32))) {
             x_error("%s: Can’t wait for more than %us", X_STRFUNC, X_MAXINT32);
         }
 
@@ -629,9 +866,9 @@ xboolean x_cond_wait_until(XCond *cond, XMutex *mutex, xint64 end_time)
 
         return success;
     }
-
 #undef KERNEL_SPAN_SEC_TYPE
 #endif
 
-    x_assert_not_reached ();
+    x_assert_not_reached();
 }
+#endif